+++ backend/models/engine/db_storage.py
@@ -5,7 +5,7 @@ from os import getenv
 from sqlalchemy import create_engine
 from sqlalchemy.orm import scoped_session, sessionmaker
 import models
-from models.base_models import BaseModel, Base
+from models.base_model import BaseModel, Base
 from models.category import Category
 from models.user import User
 from models.expense import Expense
