+++ backend/requirements.txt
@@ -8,6 +8,8 @@ greenlet==3.0.3
 itsdangerous==2.1.2
 Jinja2==3.1.3
 MarkupSafe==2.1.5
+mysqlclient==2.2.4
+pep8==1.7.1
 pycodestyle==2.11.1
 pytz==2024.1
 six==1.16.0
